split to .h and .cpp, no other updates
